Move and split api benchmarks
