+++ saleor/graphql/product/tests/benchmark/test_collection.py
