+++ saleor/graphql/product/tests/benchmark/test_variant.py
