+++ saleor/graphql/product/tests/benchmark/test_variant_stocks.py
